Replaced the result datatype with option
The datatype result was introducted to avoid the awkward use case
of using SOME("") to indicate success for operations that produce
no output. On second thought, distinguishing between two types of
success, that with and without text output, produces little
benefit for the price of being rather puzzling. The type string
option singals the intent more clearly.