feat!: print nothing if propery is missing
* BREAKING CHANGE: no longer print the section name on
`get existing-section missing-property`.
It is not clear that it helps users in any way.
It might actually be confusing, creating an impression
that the section is empty when trying to convey the opposite.
This change only affects getting a property.

Do not print a blank line when there is nothing to output from `get`
or `delete`.